refactor: extract settings screen logic into dedicated feature module
- Move settings implementation from tab page to feature module
- Simplify settings.tsx to just import and render SettingScreen
- Update TabScreenWrapper default padding to 0 for better component control
- Consolidate header padding styling in TabScreenWrapper
